Final commit; ready for review
Added styles to the components of the form, such as including a color when the user clicks on the field/input area/ et cetera. Also used regex for the phone number and postal code. Included a submit/reset button and added styles to make it bigger and appear at the bottom of the surveys.